query delete friend

+++ api/v1/friends.php
@@ -252,16 +252,16 @@ $app->delete('/friend/:id', function($id) use ($app) {
 		$userID=$session['userID'];
 		$friendID=$id;
 		//look up if friend exists
-		$sel_friends = $db->preparedStmt("SELECT f.userID,f.friendID,f.since FROM friends AS f WHERE f.userID = ? AND f.friendID = ?");
-		$sel_friends->bind_param('ii', $userID, $friendID);
+		$sel_friends = $db->preparedStmt("SELECT f.userID,f.friendID,f.since FROM friends AS f WHERE f.userID = ? AND f.friendID = ? OR f.userID = ? AND f.friendID = ? ");
+		$sel_friends->bind_param('iiii', $userID, $friendID ,$friendID ,$userID);
 		$sel_friends->execute();
 		$sel_friends->store_result();
 		
 		//delete if exists
 		if($sel_friends->num_rows>0){
 		
-			$del_friends = $db->preparedStmt("DELETE FROM friends WHERE f.userID = ? AND f.friendID = ?");
-			$del_friends->bind_param('ii', $userID, $friendID);
+			$del_friends = $db->preparedStmt("DELETE FROM friends WHERE f.userID = ? AND f.friendID = ? OR f.userID = ? AND f.friendID = ?");
+			$del_friends->bind_param('iiii', $userID, $friendID ,$friendID ,$userID);
 			$del_friends->execute();
 			$del_friends->store_result();	
 			$response['status'] = "success";
